+++ lib/ansible/executor/playbook_executor.py
@@ -72,7 +72,7 @@ class PlaybookExecutor:
         try:
             for playbook_path in self._playbooks:
                 pb = Playbook.load(playbook_path, variable_manager=self._variable_manager, loader=self._loader)
-                self._inventory.set_playbook_basedir(os.path.dirname(playbook_path))
+                self._inventory.set_playbook_basedir(os.path.realpath(os.path.dirname(playbook_path)))
 
                 if self._tqm is None: # we are doing a listing
                     entry = {'playbook': playbook_path}
@@ -147,7 +147,7 @@ class PlaybookExecutor:
                             result = self._tqm.run(play=play)
 
                             # break the play if the result equals the special return code
-                            if result == self._tqm.RUN_FAILED_BREAK_PLAY:
+                            if result & self._tqm.RUN_FAILED_BREAK_PLAY != 0:
                                 result = self._tqm.RUN_FAILED_HOSTS
                                 break_play = True