+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Ansible Changes By Release
 
 * Added the `listen` feature for modules. This feature allows tasks to more easily notify multiple handlers, as well as making it easier for handlers from decoupled roles to be notified.
 * Added support for binary modules
+* `raw` now returns `changed: true` to be consistent with shell/command/script modules. Add `changed_when: false` to `raw` tasks to restore the pre-2.2 behavior if necessary. 
 * The service module has been changed to use system specific modules if they exist and fallback to the old service module if they cannot be found or detected.
 * Several Windows facts were modified or renamed for consistency with their Unix counterparts, and many new facts were added. If your playbooks rely on any of the following keys, please ensure they are using the correct key names and/or values:
   - ansible_date_time.date (changed to use yyyy-mm-dd format instead of default system-locale format)
@@ -66,7 +67,7 @@ Ansible Changes By Release
 ###Minor Changes:
 * now -vvv shows exact path from which 'currently executing module' was picked up from.
 
-## 2.1 "The Song Remains the Same" - ACTIVE DEVELOPMENT
+## 2.1 "The Song Remains the Same"
 
 ###Major Changes:
 
@@ -75,7 +76,7 @@ Ansible Changes By Release
 * Added new modules for Azure (see below for the full list)
 * Added the ability to specify includes as "static" (either through a configuration option or on a per-include basis). When includes are static,
   they are loaded at compile time and cannot contain dynamic features like loops.
-* Added a new strategy `debug`, which allows per-task debugging of playbooks.
+* Added a new strategy `debug`, which allows per-task debugging of playbooks, for more details see https://docs.ansible.com/ansible/playbooks_debugger.html
 * Added a new option for tasks: `loop_control`. This currently only supports one option - `loop_var`, which allows a different loop variable from `item` to be used.
 * Added the ability to filter facts returned by the fact gathering setup step using the `gather_subset` option on the play or in the ansible.cfg configuration file.
   See http://docs.ansible.com/ansible/intro_configuration.html#gathering for details on the format of the option.